+++ test/test_mav2odid.c
@@ -37,11 +37,11 @@ static void print_mavlink_location(mavlink_open_drone_id_location_t *location)
            location->status, location->direction, location->speed_horizontal,
            location->speed_vertical, location->latitude, location->longitude,
            location->height_reference ? "Ground" : "TakeOff",
-           location->altitude_barometric, location->altitude_geodetic,
-           location->height,
+           (double) location->altitude_barometric,
+           (double) location->altitude_geodetic, (double) location->height,
            location->horizontal_accuracy, location->vertical_accuracy,
            location->barometer_accuracy, location->speed_accuracy,
-           location->timestamp_accuracy, location->timestamp);
+           location->timestamp_accuracy, (double) location->timestamp);
 }
 
 static void print_mavlink_auth(mavlink_open_drone_id_authentication_t *auth)
@@ -80,7 +80,7 @@ static void print_mavlink_system(mavlink_open_drone_id_system_t *system)
            system->operator_location_type, system->classification_type,
            system->operator_latitude, system->operator_longitude,
            system->area_count, system->area_radius,
-           system->area_ceiling, system->area_floor,
+           (double) system->area_ceiling, (double) system->area_floor,
            system->category_eu, system->class_eu);
 }
 
@@ -181,7 +181,7 @@ static void test_location(mav2odid_t *m2o, ODID_UAS_Data *uas_data)
         .direction = (uint16_t) (27.4f * 100),
         .speed_horizontal = (uint16_t) (4.25f * 100),
         .speed_vertical = (uint16_t) (4.5f * 100),
-        .latitude = (int32_t) (51.477f * 1E7),
+        .latitude = (int32_t) (51.477 * 1E7),
         .longitude = (int32_t) (0.0005 * 1E7),
         .altitude_barometric = 37.5f,
         .altitude_geodetic = 36.5f,
@@ -288,8 +288,8 @@ static void test_system(mav2odid_t *m2o, ODID_UAS_Data *uas_data)
     mavlink_open_drone_id_system_t system = {
         .operator_location_type = MAV_ODID_OPERATOR_LOCATION_TYPE_TAKEOFF,
         .classification_type = MAV_ODID_CLASSIFICATION_TYPE_EU,
-        .operator_latitude = (int32_t) (51.477f * 1E7),
-        .operator_longitude = (int32_t) (0.0005f * 1E7),
+        .operator_latitude = (int32_t) (51.477 * 1E7),
+        .operator_longitude = (int32_t) (0.0005 * 1E7),
         .area_count = 350,
         .area_radius = 55,
         .area_ceiling = 75.5f,
